requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
big-sleep==0.8.5
bitarray==1.2.2
boto3==1.18.15
botocore==1.21.15
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
cycler==0.10.0
cytoolz==0.11.0
einops==0.3.0
eth-abi==2.1.1
eth-account==0.5.5
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
fire==0.4.0
ftfy==6.0.3
hexbytes==0.2.1
idna==3.2
importlib-metadata==4.6.3
ipfshttpclient==0.7.0
jmespath==0.10.0
jsonschema==3.2.0
kiwisolver==1.3.1
lru-dict==1.1.7
matplotlib==3.4.3
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
numpy==1.21.1
parsimonious==0.8.1
Pillow==8.3.1
protobuf==3.17.3
pycryptodome==3.10.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-pretrained-biggan==0.1.1
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
rlp==2.0.1
s3transfer==0.5.0
six==1.16.0
termcolor==1.1.0
toolz==0.11.1
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.0
typing-extensions==3.10.0.0
urllib3==1.26.6
varint==1.0.2
wcwidth==0.2.5
web3==5.21.0
websockets==9.1
yarl==1.6.3
zipp==3.5.0
